Move RSAKeyFromPEM from package api to tokenpki
