+++ tokenpki/cert.go
@@ -49,6 +49,15 @@ func PEMRSAPrivateKey(key *rsa.PrivateKey) []byte {
 	return pem.EncodeToMemory(block)
 }
 
+// RSAKeyFromPEM decodes a PEM RSA private key.
+func RSAKeyFromPEM(key []byte) (*rsa.PrivateKey, error) {
+	block, _ := pem.Decode(key)
+	if block.Type != "RSA PRIVATE KEY" {
+		return nil, errors.New("PEM type is not RSA PRIVATE KEY")
+	}
+	return x509.ParsePKCS1PrivateKey(block.Bytes)
+}
+
 // PEMCertificate returns derBytes encoded as a PEM block.
 func PEMCertificate(derBytes []byte) []byte {
 	block := &pem.Block{
